==> tp2/Final/2-Armado de Labels.ipynb <==
import pandas as pd
import matplotlib as mp
import numpy as np
from datetime import timedelta
import datetime as dt

%matplotlib inline

#Root de archivos de datos
root_dir = '../data/'

#Lee menos data si es testing
is_testing = False

# Cantidad minima de aparicion de equipos en un dataframe
min_devices = 5

# Ventanas de tiempo
windows = pd.DataFrame({
    'begin_date': [dt.datetime(2019,4,18),dt.datetime(2019,4,21),dt.datetime(2019,4,24)],
    'window_nr':[1,2,3]
})

#Optimizado para menos memoria
auction_dtypes = {
    'device_id': np.int64,
    'source_id': np.int8
}

auctions = pd.read_csv(root_dir + 'auctions.csv.gzip',
                       compression = 'gzip',
                       dtype = auction_dtypes,
                       usecols=list(auction_dtypes.keys()) + ['date'],
                       parse_dates = ['date'])
auctions.rename({'device_id':'ref_hash',
                 'date':'created'}, inplace=True, axis='columns')
auctions['n'] = 1

# Para hacer pruebas ocupando menos memoria se hace un sampleo aleatorio de la mitad del dataframe y se elimina el resto
if is_testing:
    auctions = auctions.sample(frac=.30)
    print("Is Testing, #records:", auctions['n'].sum())
else:
    print("Not Testing, #records:", auctions['n'].sum())

installs_dtypes = {
    'application_id': np.int32,
    'ref_hash': np.int64, 
    #'click_hash':'category',
    'attributed': 'category',
    'implicit': 'category',
    #'device_countrycode': 'object', 
    'device_brand': 'object',
    'device_model': 'object', 
    'session_user_agent': 'object', 
    'user_agent': 'object', 
    'event_uuid':'object',
    'kind': 'object',
    'wifi': 'object', 
    'trans_id': 'object', 
    #'ip_address':'object', 
    'device_language': 'object'
}
install_cols = list(installs_dtypes.keys()) + ['created']
installs = pd.read_csv(root_dir + 'installs.csv.gzip', 
                       compression='gzip', 
                       usecols=install_cols,
                       dtype= installs_dtypes,
                      parse_dates=['created'])

installs['n'] = 1
print('#records:', installs['n'].sum())

unique_hashes = pd.read_csv(root_dir+'unique_hashes.csv')

---
## 2. Arreglo de los datos

### Separo las semanas de entrenamiento

Se utiliza el siguiente metodo: 

1. Ventana del 18 al 20 inclusive (1) -> Predice valores entre el 21 y 24 (2)
2. Ventana del 21 al 23 inclusive (2) -> Predice valores entre el 24 y 26 (3)

#Agrego la ventana de tiempo
auctions = auctions.loc[auctions['ref_hash'].isin(unique_hashes['ref_hash'])]\
                   .sort_values(by='created')
auctions = pd.merge_asof(auctions,windows,left_on='created',right_on='begin_date')

installs = installs.loc[installs['ref_hash'].isin(unique_hashes['ref_hash'])]
installs.sort_values(by='created',inplace=True)
installs = pd.merge_asof(installs,windows,left_on='created',right_on='begin_date')

### 2.2 Arreglo de datos de Installs

installs['kind'] = installs['kind'].str.replace(' ','_')
installs['kind'] = installs['kind'].str.replace('af_app_open ','af_app_opened')
installs['kind'] = installs['kind'].str.replace('af_app_opend','af_app_opened')
installs['kind'] = installs['kind'].str.lower()

installs.isnull().sum()

for i in ['device_brand','device_model','session_user_agent','user_agent','kind','wifi','device_language']:
    installs[i] = installs[i].fillna('unknown')
installs['device_brand'] = installs['device_brand'].astype('category')
installs['device_model'] = installs['device_model'].astype('category')
installs['session_user_agent'] = installs['session_user_agent'].astype('category')
installs['user_agent'] = installs['user_agent'].astype('category')
installs['kind'] = installs['kind'].astype('category')
installs['wifi'] = installs['wifi'].astype('category')
installs['device_language'] = installs['device_language'].astype('category')

installs.head().transpose()

---
## 3. Armado de Labels

group = ['ref_hash','window_nr']
max_dates = pd.DataFrame({'w_max_date': [dt.datetime(2019,4,21),dt.datetime(2019,4,24),dt.datetime(2019,4,27)], 'window_nr' : [1,2,3]})

sample = auctions.tail(100)

sample.head()

temp = unique_hashes
temp['window_nr'] = 1
devices = temp.copy()
#temp['window_nr'] = 2
#devices = devices.append(temp)

devices['window_nr'].value_counts()

%%time
auctions.sort_values(by=group+['created'], inplace=True)

%%time
installs.sort_values(by=group+['created'], inplace=True)

### 3.1 Label Tiempo hasta reaparicion (St)

max_secs = 3*24*60*60
temp = auctions.loc[auctions['window_nr'] != 1]\
               .groupby(group,sort=False)\
               .head(1)\
               .copy()
temp['window_nr'] = temp['window_nr'] - 1
temp = temp.set_index(group)
st = (temp['created'] - temp['begin_date']).transform(lambda x: round(x.total_seconds()))\
                                           .rename('target_st')

target_st = devices.set_index(group)\
                   .merge(st, how='left', left_index=True, right_index=True)\
                   .fillna(max_secs)

target_st.head()

### 3.1 Label Tiempo hasta install (Sc)

temp = installs.loc[installs['window_nr'] != 1]\
               .groupby(group,sort=False)\
               .head(1)\
               .copy()
temp['window_nr'] = temp['window_nr'] - 1
temp = temp.set_index(group)
sc = (temp['created'] - temp['begin_date']).transform(lambda x: round(x.total_seconds()))\
                                           .rename('target_sc')

target_sc = devices.set_index(group)\
                   .merge(sc, how='left',left_index=True, right_index=True)\
                   .fillna(max_secs)

target_sc.count()

### 3.3. Merge y guardo a file

targets = target_st.merge(target_sc, how='left',left_index=True, right_index=True)

targets.to_csv("targets.csv", header=True,index=True)
