==> tp2/Final/5-Predicciones con XGB-w all windows-bis.ipynb <==
import pandas as pd
import matplotlib as mp
import numpy as np
import datetime
import pprint
from math import sqrt

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
import xgboost as xgb
from sklearn.model_selection import GridSearchCV

%matplotlib inline

#Root de archivos de datos
root_dir = '../data/'
group=['ref_hash','window_nr']

---
## Read de Features

features = pd.read_csv('training_set6.csv',low_memory=False).set_index(group)

features.head()

features.is_last_weekend.mean()

new_features = features.groupby('ref_hash').mean()

new_features.head()

targets = pd.read_csv('targets.csv').set_index(group)

targets.head(2)

targets.isnull().sum()

training_set = features.merge(targets, how='inner', left_index=True,right_index=True)

training_set.head()

x_st = training_set[[x for x in training_set.columns if x not in targets.columns]]
y_st = training_set['target_st']

x_sc = training_set[[x for x in training_set.columns if x not in targets.columns]]
y_sc = training_set['target_sc']

---
## Training 

xtrain_st, xtest_st, ytrain_st, ytest_st = train_test_split(x_st, y_st, train_size=0.7)
xtrain_sc, xtest_sc, ytrain_sc, ytest_sc = train_test_split(x_sc, y_sc, train_size=0.7)

model_st = xgb.XGBRegressor(objective='reg:squarederror')
model_sc = xgb.XGBRegressor(objective='reg:squarederror')

model_st.fit(xtrain_st, ytrain_st)
model_sc.fit(xtrain_sc, ytrain_sc)

#def xgb_model_select(X, y):  
#    print ('Select Model...')
#    start_time  = datetime.datetime.now()
#    xgb_clf = xgb.XGBRegressor() 
#    parameters = {'objective':['reg:squarederror'],'n_estimators': [i for i in range(20,110,30)], 'max_depth':[i for i in range (3,6)]}
#    grid_search = GridSearchCV(estimator=xgb_clf, param_grid=parameters, cv=10, n_jobs=4, scoring='neg_mean_squared_error')
#    print("parameters:")
#    pprint.pprint(parameters)
#    grid_search.fit(X, y)
#    print("Best score: %0.3f" % grid_search.best_score_)
#    print("Best parameters set:")
#    best_parameters=grid_search.best_estimator_.get_params()
#    for param_name in sorted(parameters.keys()):
#        print("\t%s: %r" % (param_name, best_parameters[param_name]))
#    end_time = datetime.datetime.now()
#    print ('Select Done..., Time Cost: %d' % ((end_time - start_time).seconds) )

#xgb_model_select(x_sc,y_sc)

ypred_st = model_st.predict(xtest_st)
print("Error St.:", sqrt(mean_squared_error(ytest_st,ypred_st)))

ypred_sc = model_sc.predict(xtest_sc)
print("Error Sc.:", sqrt(mean_squared_error(ytest_sc,ypred_sc)))

#1.Feature default:
#Error St.: 83474.15593257314
#Error Sc.: 50173.18506919942
#2.Feature2
#('Error St.:', 82811.50353336304)
#('Error Sc.:', 48090.430099387835)
#3.Features3
#('Error St.:', 82312.86628044248)
#('Error Sc.:', 46461.809478466035)

model_st.feature_importances_

model_sc.feature_importances_

---
## Prediccion

# Labels a submitir con las predicciones

to_predict = pd.read_csv(root_dir + 'target_competencia_ids.csv',usecols=['ref_hash'])
to_predict.head()

X = new_features

y_st = model_st.predict(X)
y_sc = model_sc.predict(X)

results_st = X.reset_index()['ref_hash'].transform(lambda x: str(x) + '_st').rename('ref_hash').to_frame()
results_st['obj'] = y_st

results_sc = X.reset_index()['ref_hash'].transform(lambda x: str(x) + '_sc').rename('ref_hash').to_frame()
results_sc['obj'] = y_sc

results = results_st.append(results_sc)

targets_mean = targets.reset_index()
targets_mean = targets_mean.loc[~targets_mean['ref_hash'].isin(X.reset_index()['ref_hash'])].groupby('ref_hash')[['target_st','target_sc']].mean().reset_index()

t_st = targets_mean['ref_hash'].transform(lambda x: str(x) + '_st').rename('ref_hash').to_frame()
t_st['obj'] = targets_mean['target_st']

t_sc = targets_mean['ref_hash'].transform(lambda x: str(x) + '_sc').rename('ref_hash').to_frame()
t_sc['obj'] = targets_mean['target_sc']

r = t_st.append(t_sc)

results = results.append(r)

to_predict.merge(results, on='ref_hash', how='left')['obj'].count()

#Armado de File Final
max_secs = 3*24*60*60
to_predict.merge(results, on='ref_hash', how='left').fillna(max_secs).to_csv('predictions_xgb6.csv', header=True, index=False)

to_predict.count()
